fix solov2 deploy

+++ ppdet/modeling/heads/solov2_head.py
@@ -508,7 +508,7 @@ class SOLOv2Head(nn.Layer):
         sum_masks = paddle.sum(seg_masks, axis=[1, 2])
 
         y = paddle.zeros(shape=paddle.shape(sum_masks), dtype='float32')
-        keep = paddle.where(sum_masks > strides, sum_masks, y)
+        keep = paddle.where(sum_masks > strides.cast(sum_masks.dtype), sum_masks, y)
         keep = paddle.nonzero(keep)
         keep = paddle.squeeze(keep, axis=[1])
         # Prevent empty and increase fake data
